Expand gh list from 30 to 90000
It seems a bit high and the relevant entries are always on the top anyways.
So we only need the latest x new entries since this script was run last.
For the first run I would like to have all. Which also would work with
9k instead of 90k but whatever. I will leave this in for now.

But if this is causing performance issues at some point feel free to
trim it down to something small like 500.